Move two acceptance test files to the new test acc directory
